Add mitogen_host variable as optional target
The option `mitogen_host` has been added allowing users to define it as
a given target which is useful in inventories where the `ansible_host`
is already defined and may be an IP address.

> This option was added to cater for the situation where containers are
  in a givne ansible inventory but are also assigned an IP address. The
  `mitogen_host` option can now be used to point mitogen at a given
  container instead of assuming `ansible_host` is the name of the
  container. This was developed with OpenStack-Ansible in mind and has
  been validated using LXC and nspawn containers which have multiple
  IP addresses.

Documentation has been updated to show how `mitogen_host` is used and
will fall back to the default: `ansible_host` and `inventory_hostname`.

osa-containers test inventory has been updated to reflect the new
option and how it is used.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>