Make OpenStack collections qualify for NewStylePlanner
To be detected as a NewStyle module mitogen expects that your module
imports ansible.module_utils. However the OpenStack collections modules
import a utility module which then imports ansible.module_utils.
Due to the naive way mitogen checks for this import the Openstack collections
modules do not qualify currently as being NewStyle as there is a layer of abstraction,
therefore causing them to fail down the line as they are run with the OldStylePlanner.
My fix is as equally naive!