mitogen: Kill (hung) bootstrap processes after 5 second timeout
Since using select.select() in the first stage (to handle an obscure corner
case where stdin appears to be non-blocking) there has been a report of first
stage processes running for ever in an infinite loop - reading 0 bytes from
stdin.

This attempts to do an end run around that problem by aborting if the
bootstrap takes longer than a few seconds for *any* reason. Existing retry
logic should deal with it as before.

5 seconds is a best guess at a suitable timeout.